inventory-platform/services_fastapi/pyproject.toml
[project]
name = "inventory-services"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "sqlalchemy==2.0.30",
  "alembic==1.13.1",
  "psycopg[binary]==3.1.18",
  "pydantic-settings==2.3.3",
  "python-jose==3.3.0",
  "pytest==8.2.2",
  "httpx==0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
